remove browser and notify on linux and windows platform

+++ packages/app/lib/page/login/AccountSetting/AccountSetting_page.dart
@@ -241,44 +241,48 @@ class AccountSettingPage extends GetView<AccountSettingController> {
                           ),
                       ],
                     ),
-                    SettingsSection(
-                      tiles: [
-                        SettingsTile.switchTile(
-                          initialValue: controller.identity.value.enableBrowser,
-                          leading: const Icon(CupertinoIcons.compass),
-                          title: const Text("Browser ID"),
-                          onToggle: (value) async {
-                            if (value == false) {
-                              int count = await DBProvider.database.identitys
-                                  .filter()
-                                  .enableBrowserEqualTo(true)
-                                  .count();
-                              if (count == 1) {
-                                EasyLoading.showError(
-                                    "You cannot disable the last ID");
-                                return;
-                              }
+                    if (GetPlatform.isIOS ||
+                        GetPlatform.isAndroid ||
+                        GetPlatform.isMacOS)
+                      SettingsSection(
+                        tiles: [
+                          SettingsTile.switchTile(
+                            initialValue:
+                                controller.identity.value.enableBrowser,
+                            leading: const Icon(CupertinoIcons.compass),
+                            title: const Text("Browser ID"),
+                            onToggle: (value) async {
+                              if (value == false) {
+                                int count = await DBProvider.database.identitys
+                                    .filter()
+                                    .enableBrowserEqualTo(true)
+                                    .count();
+                                if (count == 1) {
+                                  EasyLoading.showError(
+                                      "You cannot disable the last ID");
+                                  return;
+                                }
 
-                              await BrowserConnect.deleteByPubkey(
-                                  controller.identity.value.secp256k1PKHex);
-                            }
-                            controller.identity.value.enableBrowser = value;
-                            await IdentityService.instance
-                                .updateIdentity(controller.identity.value);
-                            controller.identity.refresh();
-                          },
-                        ),
-                        if (controller.identity.value.enableBrowser)
-                          SettingsTile.navigation(
-                            leading: const Icon(Icons.web),
-                            title: const Text("Logged-in Websites"),
-                            onPressed: (context) async {
-                              Get.to(() => BrowserConnectedWebsite(
-                                  controller.identity.value));
+                                await BrowserConnect.deleteByPubkey(
+                                    controller.identity.value.secp256k1PKHex);
+                              }
+                              controller.identity.value.enableBrowser = value;
+                              await IdentityService.instance
+                                  .updateIdentity(controller.identity.value);
+                              controller.identity.refresh();
                             },
-                          )
-                      ],
-                    ),
+                          ),
+                          if (controller.identity.value.enableBrowser)
+                            SettingsTile.navigation(
+                              leading: const Icon(Icons.web),
+                              title: const Text("Logged-in Websites"),
+                              onPressed: (context) async {
+                                Get.to(() => BrowserConnectedWebsite(
+                                    controller.identity.value));
+                              },
+                            )
+                        ],
+                      ),
                   ],
                 )))
           ],
